Fix bug where conditional resources were being resolved
Conditional resources were being resolved even when the condition was
not met. We now only resolve the resource if the condition is met. A
little refactoring was also done to make the code more readable and to
slightly improve performance. Future work could be done to search for
instrinsic functions instead of looping over every key, value.