allow ssm dynamic references to not include a version
I did not mean to refactor the entire function. I just didn't realize
how many edge cases there were when handling dynamic references. I used
chatGPT for the regex. It took many tries before it gave me a working
one. When I asked it why it took so long for us to derive the correct
answer, it replied with a very generic list of reasons of why it might
have taken so long. The top of that list was miscommunication. I asked
it if it thought we miscommunicated and surprisingly it was able to
clearly articulate how we miscommunicated. I am both excited and afraid
of what the next 10 years will bring.