change the threshold for codecov's patch status check
This check is based on how much coverage a PR has. The current value
was 3%, which meant that 97% of an PR had to be covered for the check to
pass. This was too high and often would stop PR's that were opened by
renovate that only changed dependnency versions and didn't modify and
project code. Changing this to 70% seems like a much more reasonable
target to hit.