fix rendering of resources with a condition but not Properties
Their Condition key was not being resolved correctly which caused the
value for the entire resource to be set as the value from the condition.
This caused the program to crash when trying to render the template.
Thank to @dhutchison for both reporting the bug and coming up with this
simple but elegant fix. See #246 for more details.

Co-authored-by: David Hutchison <dhutchison86+github@googlemail.com>