remove check for nested intrinsic functions
This feature was based on the nested function limitations in the AWS
CloudFormation documentation. However, in turns out that the AWS
documentation is incorrect and nested functions are allowed in many
places where the documentation says they are not. Since this feature
was doing more harm than good, it has been removed.